backend/requirements.txt
asgiref==3.9.0
certifi==2025.7.9
charset-normalizer==3.4.2
cloudinary==1.44.1
dj-database-url==3.0.1
Django==5.2.4
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-ratelimit==4.1.0
djangorestframework==3.16.0
gunicorn==23.0.0
idna==3.10
packaging==25.0
pillow==11.3.0
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg-pool==3.2.6
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
ulid-py==1.1.0
urllib3==2.5.0
whitenoise==6.9.0
